Updating player controller
placing the method to set casting to true within a function so I can move the net from the top to bottom on the first press. I might need to add another bool to check if the 'camera' is showing the jetty or under the sea (to avoid the player pressing it during play). After this movement, casting wil be set to true.